realtime_audio/requirements.txt
openai-whisper
numpy
sounddevice
requests
python-dotenv
groq